refactor: post 관련 타입 세분화 및 공통 타입 분리
