+++ src/constants/index.ts
@@ -16,4 +16,7 @@ export type {
 export { ROOMMATE_PREFERENCE, ROOMMATE_PREFERENCE_LABEL } from "./roommatePreference";
 export type { RoommatePreference } from "./roommatePreference";
 
+export { MEMBER_ROLE } from "./memberRole";
+export type { MemberRole } from "./memberRole";
+
 export { ROOM_SIZE_LABEL, ROOM_SIZE_MAX, SEMESTER_LABEL, DORMITORY_LABEL } from "./labels";
