+++ src/constants/memberRole.ts
@@ -1,0 +1,6 @@
+export const MEMBER_ROLE = {
+  LEADER: "LEADER",
+  MEMBER: "MEMBER",
+} as const;
+
+export type MemberRole = (typeof MEMBER_ROLE)[keyof typeof MEMBER_ROLE];
