+++ src/features/board/types/index.ts
@@ -1,0 +1,32 @@
+export type {
+  TagItem,
+  BasicTagCategory,
+  HabitCategory,
+  Habit,
+  RoomSize,
+  PostStatus,
+  Semester,
+  Dormitory,
+  Campus,
+  Recycling,
+  PhoneCall,
+  ItemSharing,
+  LightsOutTime,
+  SharedLifestyle,
+} from "@/types";
+
+export type {
+  PostAuthor,
+  PostMember,
+  PostRoommatePreference,
+  PostDetail,
+  PostListItem,
+  PostListData,
+} from "./post";
+
+export type {
+  CreatePostRequest,
+  UpdatePostRequest,
+  PostListFilterParams,
+  GetPostsParams,
+} from "./postRequest";
