+++ src/types/api.ts
@@ -1,0 +1,8 @@
+/** API 공통 응답 래퍼 */
+type ApiResponse<T> = {
+  code: number;
+  message: string;
+  data: T;
+};
+
+export type { ApiResponse };
