+++ src/types/index.ts
@@ -1,1 +1,2 @@
+export type * from "./api";
 export type * from "./common";
